feat: configure project environment
