+++ .prettierrc.json
@@ -1,0 +1,7 @@
+{
+  "trailingComma": "all",
+  "printWidth": 100,
+  "tabWidth": 2,
+  "semi": true,
+  "singleQuote": true
+}
