+++ nodemon.json
@@ -1,0 +1,7 @@
+{
+  "verbose": true,
+  "ignore": ["node_modules"],
+  "watch": ["src"],
+  "ext": "ts",
+  "exec": "ts-node"
+}
